Changed variable type
Changed variable type to short, the type of "struct kevent->filter"

+++ ktrigger.h
@@ -7,7 +7,7 @@ struct config_s {
 	char*	dir;
 	char*	command;
 	bool	loop;
-	int	filter;
+	short	filter;
 };
 
 void usage(char*);
